move random circuit generation from utils to scripts
